MSVS 2015 CTP5 Win64 warning fixes - explicit casts, remove unused list.c++

+++ c++/src/capnp/encoding-test.c++
@@ -1595,7 +1595,7 @@ TEST(Encoding, NameAnnotation) {
   EXPECT_EQ(true, root.getGoodFieldName());
   EXPECT_TRUE(root.isGoodFieldName());
 
-  root.setBar(0xff);
+  root.setBar(static_cast<int8_t>(0xff));
   EXPECT_FALSE(root.isGoodFieldName());
 
   root.setAnotherGoodFieldName(test::RenamedStruct::RenamedEnum::QUX);
